Big restructure ===============
Move the preamble outside of each section to a shared folder and
switch from using include to import for preamble files. This was
necessary as include didn't like accepting files outside the root
directory of the main tex file.

Rename main.pdf to something more useful, in this case the title of
the section.

Thats about it.